services/api/pyproject.toml
[project]
name = "virtual-stylist-api"
version = "0.1.0"
description = "Virtual Stylist API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pgvector>=0.3.6",
    "arq>=0.26",
    "redis>=5.0",
    "httpx>=0.27",
    "python-jose[cryptography]>=3.3",
    "python-multipart>=0.0.12",
    "pillow>=10.4",
    "structlog>=24.4",
    "anthropic>=0.39",
    "google-cloud-storage>=2.18",
    "slowapi>=0.1.9",
    "replicate>=1.0.7",
    "pillow-heif>=1.3.0",
    "numpy>=2.4.4",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
    "mypy>=1.13",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "psycopg[binary]>=3.2",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = [".", "../.."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
